requirements.txt
pillow~=8.1.2
click~=8.0.1
exif~=1.3.0